drop the duplicated hash to curve, consensus state and bitlist padding, and split the ecdsa path out of the crate root
